% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_arc.R
\name{add_arc}
\alias{add_arc}
\title{Add arc}
\usage{
add_arc(map, data = get_map_data(map), lat_from, lon_from, lat_to, lon_to,
  stroke_from = NULL, stroke_to = NULL, stroke_width = NULL,
  layer_id = NULL, digits = 6, palette = viridisLite::viridis)
}
\description{
Add arc
}
\examples{
\dontrun{

key <- "MAPBOX_KEY"

url <- 'https://raw.githubusercontent.com/plotly/datasets/master/2011_february_aa_flight_paths.csv'
flights <- read.csv(url)
flights$id <- seq_len(nrow(flights))
flights$stroke <- sample(1:3, size = nrow(flights), replace = T)

mapdeck( token = key, style = 'mapbox://styles/mapbox/dark-v9', pitch = 45 ) \%>\%
	add_arc(
		data = flights
		, lat_from = "start_lat"
		, lon_from = "start_lon"
		, lat_to = "end_lat"
		, lon_to = "end_lon"
		, stroke_from = "airport1"
		, stroke_to = "airport2"
		, stroke_width = "stroke"
	)
}

}
