% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_bitmap.R
\name{add_bitmap}
\alias{add_bitmap}
\title{Add bitmap}
\usage{
add_bitmap(map, url, bounds, desaturate = 0, transparent_colour = c(0,
  0, 0, 0), tint_colour = c(255, 255, 255), layer_id = NULL)
}
\arguments{
\item{bounds}{coordinates of the bounding box of the bitmap [left, bottom, right, top]}

\item{desaturate}{the desatruation of the bitmap, in range [0,1], 0 being the original colour
and 1 being greyscale}

\item{transparent_colour}{the colour to use for transparent pixels}

\item{tint_colour}{}
}
\description{
Add bitmap
}
\examples{
\donttest{

mapdeck(
  location = c(-71.516, 37.936), zoom = 12
) \%>\%
 add_bitmap(
   url = "https://docs.mapbox.com/mapbox-gl-js/assets/radar.gif"
   , bounds = list(c(-71.516, 37.936), c(-80.425, 37.936), c(-80.425, 46.437), c(-71.516, 46.437))
 )

}

}
