% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_tile3d.R
\name{add_cesium}
\alias{add_cesium}
\title{Add Cesium}
\usage{
add_cesium(map, data, layer_id = NULL, ion_token = NULL)
}
\description{
Add Cesium
}
\examples{
\donttest{

## Melbourne point cloud
ion_asset <- 43978
tile_data <- paste0("https://assets.cesium.com/",ion_asset,"/tileset.json")

mapdeck() \%>\%
 add_cesium(
   data = tile_data
   , ion_token = secret::get_secret("ion")
 )

}

}
