% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_contour.R
\name{add_contour}
\alias{add_contour}
\title{Add Contour}
\usage{
add_contour(map, data = get_map_data(map), lon = NULL, lat = NULL,
  polyline = NULL, cell_size = 1000, layer_id = NULL, digits = 6)
}
\arguments{
\item{map}{a mapdeck map object}

\item{data}{data to be used in the layer. All coordinates are expected to be in
Web Mercator Projection}

\item{lon}{column containing longitude values}

\item{lat}{column containing latitude values}

\item{polyline}{optional column of \code{data} containing the polylines, if using encoded polylines}

\item{cell_size}{size of each cell in meters}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}
}
\description{
Add Contour
}
\details{
\code{add_grid} supports POINT and MULTIPOINT sf objects
}
\examples{
\donttest{
## You need a valid access token from Mapbox
key <- 'abc'

df <- read.csv(paste0(
'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/',
'examples/3d-heatmap/heatmap-data.csv'
))

df <- df[ !is.na(df$lng ), ]

mapdeck( token = key, style = mapdeck_style("dark"), pitch = 45 ) \%>\%
add_contour(
  data = df
  , lat = "lat"
  , lon = "lng"
  , cell_size = 5000
)

## using sf object
library(sf)
sf <- sf::st_as_sf( df, coords = c("lng", "lat"))

mapdeck( token = key, style = mapdeck_style("dark"), pitch = 45 ) \%>\%
add_contour(
  data = sf
  , cell_size = 5000
)


}

}
