% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{add_dependencies}
\alias{add_dependencies}
\title{Add Dependencies}
\usage{
add_dependencies(map)
}
\description{
Javascript dependencies for mapdeck and deck.gl, useful for when not using
a Mapbox map
}
\examples{
\donttest{

## use with a google map from googleway
library(googleway)

set_key("GOOGLE_MAP_KEY")

google_map() \%>\%
	add_dependencies() \%>\%
	add_scatterplot(
		data = capitals
		, lon = "lon"
		, lat = "lat"
		, fill_colour = "country"
		, radius = 10000
	)

}



}
