% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_tile3d.R
\name{add_i3s}
\alias{add_i3s}
\title{Add I3S}
\usage{
add_i3s(map, data, layer_id = NULL)
}
\description{
Add I3S
}
\examples{
\donttest{

## San Francisco buildings
i3s <- 'https://tiles.arcgis.com/tiles/z2tnIkrLQ2BRzr6P/arcgis/rest/services/SanFrancisco_Bldgs/SceneServer/layers/0'
mapdeck() \%>\%
  add_i3s(
    data = i3s
  )
}

}
