% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_layer.R
\name{add_mapbox_source}
\alias{add_mapbox_source}
\title{Add mapbox source}
\usage{
add_mapbox_source(map, id, js)
}
\description{
Add mapbox source
}
\details{
Mapbox sources supply data to be shown on the map. The type of source is specified
by the "type" property, adn must be one of
\itemize{
  \item{vector}
  \item{raster}
  \item{raster-dem}
  \item{geojson}
  \item{image}
  \item{video}
}

See the Mapbox Sources definition at \url{https://docs.mapbox.com/mapbox-gl-js/style-spec/#sources}
}
