% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_path.R
\name{add_path}
\alias{add_path}
\title{Add Path}
\usage{
add_path(map, data = get_map_data(map), polyline, stroke_colour = NULL,
  stroke_width = NULL, digits = 6, palette = viridisLite::viridis)
}
\description{
Add Path
}
