% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_sf.R
\name{add_sf}
\alias{add_sf}
\title{Add sf}
\usage{
add_sf(map, data = get_map_data(map), ...)
}
\arguments{
\item{map}{a mapdeck map object}

\item{data}{data to be used in the layer. All coordinates are expected to be
EPSG:4326 (WGS 84) coordinate system}

\item{...}{other argumetns passed to one of the plotting layers. See details}
}
\description{
Adds an sf object to the map.
}
\details{
The plotting layer is determined by the type of sf geometries.

\itemize{
  \item{POINT and MULTIPOINT objects will call \link{add_scatterplot}}
  \item{LINESTRING and MULTILINESTRING objects will call \link{add_path}}
  \item{POLYGON and MULTIPOLYGON objects will call \link{add_polygon}}
  \item{GEOMETRY objects will call \link{add_geojson}}
}
}
