% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_terrain.R
\name{add_terrain}
\alias{add_terrain}
\title{Add terrain}
\usage{
add_terrain(map, layer_id = NULL, elevation_data, texture = NULL, bounds)
}
\arguments{
\item{map}{a mapdeck map object}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}

\item{elevation_data}{Image URL that encodes height data.}

\item{texture}{Image URL to use as the texture}

\item{elevation_decoder}{Numeric vector giving parameters to use to convert a
pixel to elevation in metres.}
}
\description{
Adds mesh surfaces from height map images
}
\examples{
\donttest{

set_token( "MAPBOX_TOKEN" )
elevation <- 'https://raw.githubusercontent.com/visgl/deck.gl-data/master/website/terrain.png'
texture <- 'https://raw.githubusercontent.com/visgl/deck.gl-data/master/website/terrain-mask.png'
bounds <- c(-122.5233, 37.6493, -122.3566, 37.8159)

mapdeck( location = bounds[1:2], zoom = 10) \%>\%
  add_terrain(
    elevation_data = elevation,
    texture = texture,
    bounds = bounds
  )

}

}
