% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_text_cluster.R
\name{add_text_cluster}
\alias{add_text_cluster}
\title{Add Text Cluster}
\usage{
add_text_cluster(map, data = get_map_data(map), text, lon = NULL,
  lat = NULL, polyline = NULL, weight = NULL, layer_id, digits = 6,
  palette = viridisLite::viridis)
}
\arguments{
\item{map}{a mapdeck map object}

\item{data}{data to be used in the layer}

\item{text}{column of \code{data} containing the text}

\item{lon}{column containing longitude values}

\item{lat}{column containing latitude values}

\item{polyline}{column of \code{data} containing the polylines}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type}

\item{digits}{integer. Use this parameter to specify how many digits (decimal places)
should be used for the latitude / longitude coordinates.}

\item{palette}{a function, or list of functions which generates hex colours}
}
\description{
The Text Layer takes in coordinate points and renders them as circles
with a certain radius.
}
\examples{

\dontrun{

## You need a valid access token from Mapbox
key <- 'abc'

df <- capitals
df$weight <- sample(1:500, size = nrow(df))

mapdeck(
  token = key,
  style = mapdeck_style('dark')
) \%>\%
  add_text_cluster(
    data = capitals
    , lon = 'lon'
    , lat = 'lat'
    , weight = 'weight'
    , text = 'capital'
    , layer_id = 'text'
  )
}

}
