% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_title.R
\name{add_title}
\alias{add_title}
\title{Add Title}
\usage{
add_title(map, title, layer_id = NULL)
}
\arguments{
\item{map}{a mapdeck map object}

\item{title}{}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}
}
\description{
Adds a title to a map
}
