% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_view_state.R
\name{add_view_state}
\alias{add_view_state}
\title{Add View State}
\usage{
add_view_state(map, layer_id = NULL)
}
\arguments{
\item{map}{a mapdeck map object}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}
}
\description{
Adds the current View State to a map. The view state consists of
\itemize{
  \item{width}
  \item{height}
  \item{latitude & longitude}
  \item{zoom}
  \item{bearing}
  \item{pitch}
  \item{altitude}
  \item{viewBounds}
  \item{interactionState}
}
}
\examples{
\donttest{

mapdeck() \%>\%
  add_view_state()
}

}
