% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capitals}
\alias{capitals}
\title{Capital cities for each country}
\format{
A data frame with 200 observations and 4 variables
\describe{
  \item{country}{country name}
  \item{capital}{capital name}
  \item{lat}{latitude of capital}
  \item{lon}{longitude of capital}
}
}
\usage{
capitals
}
\description{
A data set containing the coordinates of 200 capitical cities in the world
}
\keyword{datasets}
