% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{city_trail}
\alias{city_trail}
\title{city_trail}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 3 columns.
}
\usage{
city_trail
}
\description{
An sf object of a cyclist cycling around Melbourne's Capital City Trail
}
\keyword{datasets}
