% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_arc.R, R/map_layer_geojson.R,
%   R/map_layer_grid.R, R/map_layer_line.R, R/map_layer_path.R,
%   R/map_layer_pointcloud.R, R/map_layer_polygon.R, R/map_layer_scatterplot.R,
%   R/map_layer_screengrid.R, R/map_layer_text.R
\name{clear_arc}
\alias{clear_arc}
\alias{clear_geojson}
\alias{clear_grid}
\alias{clear_line}
\alias{clear_path}
\alias{clear_pointcloud}
\alias{clear_polygon}
\alias{clear_scatterplot}
\alias{clear_screengrid}
\alias{clear_text}
\title{Clear Arc}
\usage{
clear_arc(map, layer_id = NULL)

clear_geojson(map, layer_id = NULL)

clear_grid(map, layer_id = NULL)

clear_line(map, layer_id = NULL)

clear_path(map, layer_id = NULL)

clear_pointcloud(map, layer_id = NULL)

clear_polygon(map, layer_id = NULL)

clear_scatterplot(map, layer_id = NULL)

clear_screengrid(map, layer_id = NULL)

clear_text(map, layer_id = NULL)
}
\arguments{
\item{map}{a mapdeck map object}

\item{layer_id}{the layer_id of the layer you want to clear}
}
\description{
Clears elements from a map
}
