% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{clear_legend}
\alias{clear_legend}
\title{Clear Legend}
\usage{
clear_legend(map, layer_id)
}
\arguments{
\item{map}{the map from which you want to clear the legend.}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}
}
\description{
Clears the legend for a given layer_id
}
