% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{garmin}
\alias{garmin}
\title{garmin}
\format{An object of class \code{sf} (inherits from \code{data.table}, \code{data.frame}) with 151 rows and 3 columns.}
\usage{
garmin
}
\description{
An sf object of my cycling activities from my Garmin
}
\keyword{datasets}
