% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_light_settings.R
\name{light_settings}
\alias{light_settings}
\title{Light Settings}
\description{
List object containg light settings.
}
\details{
Available in \link{add_geojson}, \link{add_pointcloud} and \link{add_polygon}

\itemize{
  \item{numberOfLights - the number of lights. Maximum of 5}
  \item{lightsPosition - vector of x, y, z coordinates. Must be 3x the nubmer of lights}
  \item{ambientRatio - the ambient ratio of the lights}
}
}
\examples{

light <- list(
  lightsPosition = c(-150, 75, 0)
  , numberOfLights = 1
  , ambientRatio = 0.2
)

}
