% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox_map_utilities.R
\name{mapbox_dispatch}
\alias{mapbox_dispatch}
\title{mapbox dispatch}
\usage{
mapbox_dispatch(map, funcName, mapbox = stop(paste(funcName,
  "requires a map update object")), mapbox_update = stop(paste(funcName,
  "does not support map update objects")))
}
\arguments{
\item{map}{a map object, as returned from \code{\link{mapdeck}}}

\item{funcName}{the name of the function that the user called that caused
this \code{mapdeck_dispatch} call; for error message purposes}

\item{mapdeck}{an action to be performed if the map is from
\code{\link{mapdeck}}}

\item{mapdeck_update}{an action to be performed if the map is from
\code{\link{mapdeck_update}}}
}
\value{
\code{mapdeck_dispatch} returns the value of \code{mapdeck} or
or an error. \code{invokeMethod} returns the
\code{map} object that was passed in, possibly modified.
}
\description{
Extension points for plugins
}
