% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{mapdeck_legend}
\alias{mapdeck_legend}
\title{Mapdeck Legend}
\usage{
mapdeck_legend(legend_elements)
}
\arguments{
\item{legend_elements}{vector of legend elements (made from \link{legend_element})}
}
\description{
Mapdeck Legend
}
