% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{mapdeck_map}
\alias{mapdeck_map}
\title{mapdeck Map}
\usage{
mapdeck_map(key, data = NULL, width = NULL, height = NULL,
  elementId = NULL, padding = 0,
  style = "mapbox://styles/mapbox/streets-v9")
}
\description{
mapdeck Map
}
