% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_styles.R
\name{mapdeck_style}
\alias{mapdeck_style}
\title{Mapdeck Style}
\usage{
mapdeck_style(
  style = c("dark", "light", "outdoors", "streets", "satellite", "satellite-streets")
)
}
\arguments{
\item{style}{one of streets, outdoors, light, dark, satellite, satellite-streets}
}
\description{
Various styles available to all Mapbox accounts using a valid access token.
Available styles are listed at
\url{https://docs.mapbox.com/api/maps/#styles}.
}
\examples{
\donttest{

## You need a valid access token from Mapbox
key <- 'abc'

## set a map style
mapdeck(token = key, style = mapdeck_style("dark"))

}

}
