% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{mapdeck_view}
\alias{mapdeck_view}
\title{Mapdeck view}
\usage{
mapdeck_view(map, location, zoom = 6, duration = 0,
  transition = c("linear", "fly"))
}
\arguments{
\item{duration}{time in milliseconds}
}
\description{
Mapdeck view
}
