% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{md_map}
\alias{md_map}
\title{md Map}
\usage{
md_map(key, data = NULL, width = NULL, height = NULL, elementId = NULL,
  padding = 0, style = "mapbox://styles/mapbox/streets-v9", pitch = 0,
  location = c(0, 0))
}
\arguments{
\item{location}{vector of lon and lat coordinates}
}
\description{
md Map
}
