% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melbourne}
\alias{melbourne}
\title{Polygons in and around Melbourne}
\format{
An sfencoded and data frame object with 41 observations and 8 variables.
See library googlePolylines for information on sfencoded objects
}
\usage{
melbourne
}
\description{
A data set containing statistical area 2 regions of central (and surrounds) Melbourne.
}
\keyword{datasets}
