% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{melbourne_mesh}
\alias{melbourne_mesh}
\title{Melbourne Mesh}
\format{
An object of class \code{mesh3d} (inherits from \code{shape3d}) of length 6.
}
\usage{
melbourne_mesh
}
\description{
A mesh3d object of Melbourne
}
\keyword{datasets}
