% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{road_safety}
\alias{road_safety}
\title{road_safety}
\format{
An object of class \code{data.frame} with 19139 rows and 2 columns.
}
\usage{
road_safety
}
\description{
A data.frame of counts of traffic accidents in the UK
}
\keyword{datasets}
