% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roads}
\alias{roads}
\title{Roads in central Melbourne}
\format{
An sf and data frame object with 18286 observations and 16 variables
}
\usage{
roads
}
\description{
A simple feature \code{sf} object of roads in central Melbourne
}
\details{
Obtained from \url{https://www.data.gov.au} and distributed under the Creative Commons 4 License
\url{https://creativecommons.org/licenses/by/4.0/}
}
\keyword{datasets}
