% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_tokens.R
\name{set_token}
\alias{set_token}
\title{Set Token}
\usage{
set_token(token)
}
\arguments{
\item{token}{Mapbox access token}
}
\description{
Sets an access token so it's available for all mapdeck calls. See details
}
\details{
Use \code{set_token} to make access tokens available for all the \code{mapdeck()}
calls in a session so you don't have to keep specifying the \code{token} argument
each time
}
