% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{update_style}
\alias{update_style}
\title{update style}
\usage{
update_style(map, style)
}
\arguments{
\item{map}{a mapdeck map object}

\item{style}{the style of the map (see \link{mapdeck_style})}
}
\description{
update style
}
