% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewdeck.R
\name{viewdeck_dispatch}
\alias{viewdeck_dispatch}
\title{viewdeck dispatch}
\usage{
viewdeck_dispatch(map, funcName, viewdeck = stop(paste(funcName,
  "requires a map update object")),
  viewdeck_update = stop(paste(funcName,
  "does not support map update objects")))
}
\arguments{
\item{map}{a map object, as returned from \code{\link{viewdeck}}}

\item{funcName}{the name of the function that the user called that caused
this \code{mapdeck_dispatch} call; for error message purposes}

\item{viewdeck}{an action to be performed if the map is from
\code{\link{viewdeck}}}

\item{mapdeck_update}{an action to be performed if the map is from
\code{\link{mapdeck_update}}}
}
\value{
\code{mapdeck_dispatch} returns the value of \code{viewdeck} or
or an error. \code{invokeMethod} returns the
\code{map} object that was passed in, possibly modified.
}
\description{
Extension points for plugins
}
