% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_arc.R
\name{add_arc}
\alias{add_arc}
\title{Add arc}
\usage{
add_arc(map, data = get_map_data(map), layer_id, lat_from, lon_from, lat_to,
  lon_to, id = NULL, stroke_from = NULL, stroke_to = NULL,
  stroke_width = NULL, digits = 6, palette = viridisLite::viridis)
}
\arguments{
\item{map}{a mapdeck map object}

\item{data}{data to be used in the layer}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type}

\item{lat_from}{latitude coordinate of the origin}

\item{lon_from}{longitude coordinate of the origin}

\item{lat_to}{latitude coordinate of the destination}

\item{lon_to}{latitude longitude of the destination}

\item{id}{an id value in \code{data} to identify layers when interacting in Shiny apps}

\item{stroke_from}{variable or hex colour to use as the staring stroke colour}

\item{stroke_to}{variable or hex colour to use as the ending stroke colour}

\item{stroke_width}{width of the stroke}

\item{digits}{integer. Use this parameter to specify how many digits (decimal places)
should be used for the latitude / longitude coordinates.}

\item{palette}{a function, or list of functions which generates hex colours}
}
\description{
Add arc
}
\examples{
\dontrun{

key <- "MAPBOX_KEY"

url <- 'https://raw.githubusercontent.com/plotly/datasets/master/2011_february_aa_flight_paths.csv'
flights <- read.csv(url)
flights$id <- seq_len(nrow(flights))
flights$stroke <- sample(1:3, size = nrow(flights), replace = T)

mapdeck( token = key, style = 'mapbox://styles/mapbox/dark-v9', pitch = 45 ) \%>\%
	add_arc(
		data = flights
		, lat_from = "start_lat"
		, lon_from = "start_lon"
		, lat_to = "end_lat"
		, lon_to = "end_lon"
		, stroke_from = "airport1"
		, stroke_to = "airport2"
		, stroke_width = "stroke"
	)
}

}
