% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_dependencies.R
\name{add_dependencies}
\alias{add_dependencies}
\title{Add Dependencies}
\usage{
add_dependencies(map)
}
\arguments{
\item{map}{the map object to which dependencies will be added}
}
\description{
Adds the required mapdeck javascript dependencies to a map when not using a mapdeck map.
}
\examples{
\donttest{

## use with a google map from googleway
library(googleway)

set_key("GOOGLE_MAP_KEY")

google_map() \%>\%
	add_dependencies() \%>\%
	add_scatterplot(
		data = capitals
		, lon = "lon"
		, lat = "lat"
		, fill_colour = "country"
		, radius = 10000
	)

}



}
