% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_geojson.R
\name{add_geojson}
\alias{add_geojson}
\title{Add Geojson}
\usage{
add_geojson(map, data = get_map_data(map), layer_id = NULL)
}
\description{
Add Geojson
}
\details{
the GeoJSON string needs to have a \code{class} attribute of 'json'
}
