% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_hexagon.R
\name{add_hexagon}
\alias{add_hexagon}
\title{Add hexagon}
\usage{
add_hexagon(map, data = get_map_data(map), polyline = NULL,
  lon = NULL, lat = NULL, layer_id = NULL, id = NULL,
  radius = 1000, elevation_scale = 1, auto_highlight = FALSE,
  highlight_colour = "#AAFFFFFF",
  colour_range = colourvalues::colour_values(1:6, palette = "viridis"))
}
\arguments{
\item{map}{a mapdeck map object}

\item{data}{data to be used in the layer}

\item{lon}{column containing longitude values}

\item{lat}{column containing latitude values}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}

\item{id}{an id value in \code{data} to identify layers when interacting in Shiny apps.}

\item{radius}{in metres}

\item{elevation_scale}{value to sacle the elevations of the hexagons}

\item{auto_highlight}{logical indicating if the shape under the mouse should auto-highlight}

\item{highlight_colour}{hex string colour to use for highlighting. Must contain the alpha component.}

\item{colour_range}{palette of colours}
}
\description{
The Hexagon Layer renders a hexagon heatmap based on an array of points.
It takes the radius of hexagon bin, projects points into hexagon bins.
The color and height of the hexagon is scaled by number of points it contains.
}
\details{
\code{add_hexagon} supports POINT and MULTIPOINT sf objects
}
\examples{
\dontrun{

## You need a valid access token from Mapbox
key <- 'abc'

df <- read.csv(paste0(
'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/'
, '3d-heatmap/heatmap-data.csv'
))

df <- df[!is.na(df$lng), ]

mapdeck( token = key, style = 'mapbox://styles/mapbox/dark-v9', pitch = 45 ) \%>\%
add_hexagon(
  data = df
  , lat = "lat"
  , lon = "lng"
  , layer_id = "hex_layer"
  , elevation_scale = 100
)

library( sf )
sf <- sf::st_as_sf( df, coords = c("lng", "lat"))
mapdeck( token = key, style = 'mapbox://styles/mapbox/dark-v9', pitch = 45 ) \%>\%
add_hexagon(
  data = sf
  , layer_id = "hex_layer"
  , elevation_scale = 100
)

}

}
