% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_pointcloud.R
\name{add_pointcloud}
\alias{add_pointcloud}
\title{Add Pointcloud}
\usage{
add_pointcloud(map, data = get_map_data(map), lon = NULL, lat = NULL,
  elevation = NULL, polyline = NULL, radius = 10,
  fill_colour = NULL, fill_opacity = NULL, tooltip = NULL,
  auto_highlight = FALSE, highlight_colour = "#AAFFFFFF",
  light_settings = list(), layer_id = NULL, id = NULL,
  palette = "viridis", na_colour = "#808080FF", legend = FALSE,
  legend_options = NULL, update_view = TRUE, focus_layer = FALSE)
}
\arguments{
\item{map}{a mapdeck map object}

\item{data}{data to be used in the layer}

\item{lon}{column containing longitude values}

\item{lat}{column containing latitude values}

\item{elevation}{column containing the elevation values}

\item{polyline}{column of \code{data} containing the polylines}

\item{radius}{value in pixels of each point}

\item{fill_colour}{column of \code{data} or hex colour for the fill colour}

\item{fill_opacity}{value between 1 and 255. Either a string specifying the
column of \code{data} containing the fill opacity of each shape, or a value
between 1 and 255 to be applied to all the shapes}

\item{tooltip}{variable of \code{data} containing text or HTML to render as a tooltip}

\item{auto_highlight}{logical indicating if the shape under the mouse should auto-highlight}

\item{highlight_colour}{hex string colour to use for highlighting. Must contain the alpha component.}

\item{light_settings}{list of light setting parameters. See \link{light_settings}}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}

\item{id}{an id value in \code{data} to identify layers when interacting in Shiny apps.}

\item{palette}{string or matrix. String is either one of "viridis","inferno",
"magma","plasma" or "cividis". A matrix is a 3 or 4 column numeric matrix of values
between [0, 255], where the 4th column represents the alpha.}

\item{na_colour}{hex string colour to use for NA values}

\item{legend}{either a logical indiciating if the legend(s) should be displayed, or
a named list indicating which colour attributes should be included in the legend.}

\item{legend_options}{A list of options for controlling the legend.}
}
\description{
The Pointcloud Layer takes in coordinate points and renders them as circles
with a certain radius.
}
\details{
\code{add_pointcloud} supports POINT and MULTIPOINT sf objects
}
\section{legend}{


The \code{legend_options} can be used to control the appearance of the legend.
This should be a named list, where the names are one of
\itemize{
  \item{css - a string of valid \code{css} for controlling the appearance of the legend}
  \item{title - a string to use for the title of the legend}
}

If the layer allows different fill and stroke colours, you can use different options for each. See examples in \link{add_arc}.
}

\section{id}{


The \code{id} is returned to your R session from an interactive shiny environment
by observing layer clicks. This is useful for returning the data.frame row relating to the
cliked shape.

From within a shiny server you would typically use \code{ observeEvent({input$map_arc_click})},
where 'map' is the map_id supplied to \code{mapdeckOutput()}, and 'arc' is the layer
you are clicking on
}

\examples{
\donttest{
## You need a valid access token from Mapbox
key <- 'abc'

df <- capitals
df$z <- sample(10000:10000000, size = nrow(df))

mapdeck(token = key, style = 'mapbox://styles/mapbox/dark-v9') \%>\%
add_pointcloud(
  data = df
  , lon = 'lon'
  , lat = 'lat'
  , elevation = 'z'
  , layer_id = 'point'
  , fill_colour = "country"
  , tooltip = "country"
)

## as an sf object wtih a Z attribute
library(sf)
sf <- sf::st_as_sf( df , coords = c("lon","lat","z"))

mapdeck(token = key, style = 'mapbox://styles/mapbox/dark-v9') \%>\%
add_pointcloud(
  data = sf
  , layer_id = 'point'
  , fill_colour = "country"
  , tooltip = "country"
)

}

}
