% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_pointcloud.R
\name{add_pointcloud}
\alias{add_pointcloud}
\title{Add Pointcloud}
\usage{
add_pointcloud(map, data = get_map_data(map), lon = NULL, lat = NULL,
  elevation = NULL, polyline = NULL, radius = NULL,
  fill_colour = NULL, fill_opacity = NULL, tooltip = NULL,
  light_settings = list(), layer_id = NULL, digits = 6,
  palette = "viridis", na_colour = "#808080FF", legend = FALSE,
  legend_options = NULL)
}
\arguments{
\item{map}{a mapdeck map object}

\item{data}{data to be used in the layer}

\item{lon}{column containing longitude values}

\item{lat}{column containing latitude values}

\item{elevation}{column containing the elevation values}

\item{polyline}{column of \code{data} containing the polylines}

\item{radius}{in metres}

\item{fill_colour}{column of \code{data} or hex colour for the fill colour}

\item{fill_opacity}{value between 1 and 255. Either a string specifying the
column of \code{data} containing the fill opacity of each shape, or a value
between 1 and 255 to be applied to all the shapes}

\item{tooltip}{variable of \code{data} containing text or HTML to render as a tooltip}

\item{light_settings}{list of light setting parameters. See \link{light_settings}}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}

\item{digits}{integer. Use this parameter to specify how many digits (decimal places)
should be used for the latitude / longitude coordinates.}

\item{palette}{string or matrix. String is either one of "viridis","inferno",
"magma","plasma" or "cividis". A matrix is a 3 or 4 column numeric matrix of values
between [0, 255], where the 4th column represents the alpha.}

\item{na_colour}{hex string colour to use for NA values}

\item{legend}{either a logical indiciating if the legend(s) should be displayed, or
a named list indicating which colour attributes should be included in the legend.}

\item{legend_options}{A list of options for controlling the legend.}
}
\description{
The Pointcloud Layer takes in coordinate points and renders them as circles
with a certain radius.
}
\examples{
\donttest{
## You need a valid access token from Mapbox
key <- 'abc'

df <- capitals
df$z <- sample(10000:10000000, size = nrow(df))

mapdeck(token = key, style = 'mapbox://styles/mapbox/dark-v9') \%>\%
add_pointcloud(
  data = df
  , lon = 'lon'
  , lat = 'lat'
  , elevation = 'z'
  , layer_id = 'point'
  , fill_colour = "country"
  , tooltip = "country"
)
}

}
