% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_polygon.R
\name{add_polygon}
\alias{add_polygon}
\title{Add Polygon}
\usage{
add_polygon(map, data = get_map_data(map), polyline = NULL,
  stroke_colour = NULL, stroke_width = NULL, fill_colour = NULL,
  fill_opacity = NULL, elevation = NULL, tooltip = NULL,
  auto_highlight = FALSE, highlight_colour = "#AAFFFFFF",
  light_settings = list(), layer_id = NULL, id = NULL,
  palette = "viridis", na_colour = "#808080FF", legend = FALSE,
  legend_options = NULL)
}
\arguments{
\item{map}{a mapdeck map object}

\item{data}{data to be used in the layer}

\item{polyline}{column of \code{data} containing the polylines}

\item{stroke_colour}{variable of \code{data} or hex colour for the stroke}

\item{stroke_width}{width of the stroke}

\item{fill_colour}{column of \code{data} or hex colour for the fill colour}

\item{fill_opacity}{value between 1 and 255. Either a string specifying the
column of \code{data} containing the fill opacity of each shape, or a value
between 1 and 255 to be applied to all the shapes}

\item{elevation}{the height of the polygon}

\item{tooltip}{variable of \code{data} containing text or HTML to render as a tooltip}

\item{auto_highlight}{logical indicating if the shape under the mouse should auto-highlight}

\item{highlight_colour}{hex string colour to use for highlighting. Must contain the alpha component.}

\item{light_settings}{list of light setting parameters. See \link{light_settings}}

\item{layer_id}{single value specifying an id for the layer. Use this value to
distinguish between shape layers of the same type. Layers with the same id are likely
to conflict and not plot correctly}

\item{id}{an id value in \code{data} to identify layers when interacting in Shiny apps.}

\item{palette}{string or matrix. String is either one of "viridis","inferno",
"magma","plasma" or "cividis". A matrix is a 3 or 4 column numeric matrix of values
between [0, 255], where the 4th column represents the alpha.}

\item{legend}{either a logical indiciating if the legend(s) should be displayed, or
a named list indicating which colour attributes should be included in the legend.}

\item{legend_options}{A list of options for controlling the legend.}
}
\description{
The Polygon Layer renders filled and/or stroked polygons. If using \code{sf} objects
only POLYGONs are supported, MULTIPOLYGONs are ignored.
}
\details{
\code{add_polygon} supports POLYGON and MULTIPOLYGON sf objects
}
\section{legend}{


The \code{legend_options} can be used to control the appearance of the legend.
This should be a named list, where the names are one of
\itemize{
  \item{css - a string of valid \code{css} for controlling the appearance of the legend}
  \item{title - a string to use for the title of the legend}
}

If the layer allows different fill and stroke colours, you can use different options for each. See examples in \link{add_arc}.
}

\section{id}{


The \code{id} is returned to your R session from an interactive shiny environment
by observing layer clicks. This is useful for returning the data.frame row relating to the
cliked shape.

From within a shiny server you would typically use \code{ observeEvent({input$map_arc_click})},
where 'map' is the map_id supplied to \code{mapdeckOutput()}, and 'arc' is the layer
you are clicking on
}

\examples{
\donttest{

## You need a valid access token from Mapbox
key <- 'abc'

df <- melbourne
df$elevation <- sample(100:5000, size = nrow(df))
df$info <- paste0("<b>SA2 - </b><br>",df$SA2_NAME)

mapdeck(
  token = key
  , style = mapdeck_style('dark')
  , location = c(145, -38)
  , zoom = 8
  ) \%>\%
  add_polygon(
    data = df
    , polyline = "geometry"
    , layer = "polygon_layer"
    , fill_colour = "SA2_NAME",
    , elevation = "elevation"
    , stroke_width = 0
    , tooltip = 'info'
    , legend = T
  )

library(sf)
library(geojsonsf)

sf <- geojson_sf("https://symbolixau.github.io/data/geojson/SA2_2016_VIC.json")

mapdeck(
  token = key
  , style = mapdeck_style('dark')
) \%>\%
  add_polygon(
    data = sf
    , layer = "polygon_layer"
    , fill_colour = "SA2_NAME16"
  )

}

}
