% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_scatterplot.R
\name{add_scatterplot}
\alias{add_scatterplot}
\title{Add Scatterplot}
\usage{
add_scatterplot(map, data = get_map_data(map), lat, lon, radius = NULL,
  fill_colour = NULL, fill_opacity = NULL, stroke_width = NULL,
  digits = 6, palette = viridisLite::viridis)
}
\arguments{
\item{radius}{in metres}
}
\description{
Add Scatterplot
}
