% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_tile3d.R
\name{add_tile3d}
\alias{add_tile3d}
\title{Add Tile3D}
\usage{
add_tile3d(map, data = get_map_data(map), layer_id = NULL, ion_token = NULL)
}
\description{
Add Tile3D
}
\examples{
\donttest{

canary_warf <- 'https://raw.githubusercontent.com/AnalyticalGraphicsInc/3d-tiles/master/examples/tileset.json'
mapdeck(
  style = mapdeck_style("dark")
  ) \%>\%
  add_tile3d(
    data = canary_warf
  )


}

}
