% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_animated_arc.R,
%   R/map_layer_animated_line.R, R/map_layer_arc.R, R/map_layer_bitmap.R,
%   R/map_layer_column.R, R/map_layer_geojson.R, R/map_layer_greatcircle.R,
%   R/map_layer_grid.R, R/map_layer_heatmap.R, R/map_layer_hexagon.R,
%   R/map_layer_line.R, R/map_layer_mesh.R, R/map_layer_path.R,
%   R/map_layer_pointcloud.R, R/map_layer_polygon.R, R/map_layer_scatterplot.R,
%   R/map_layer_screengrid.R, R/map_layer_terrain.R, R/map_layer_text.R,
%   R/map_layer_title.R, R/map_layer_trips.R
\name{clear_animated_arc}
\alias{clear_animated_arc}
\alias{clear_line}
\alias{clear_arc}
\alias{clear_bitmap}
\alias{clear_column}
\alias{clear_geojson}
\alias{clear_greatcircle}
\alias{clear_grid}
\alias{clear_heatmap}
\alias{clear_hexagon}
\alias{clear_mesh}
\alias{clear_path}
\alias{clear_pointcloud}
\alias{clear_polygon}
\alias{clear_scatterplot}
\alias{clear_screengrid}
\alias{clear_terrain}
\alias{clear_text}
\alias{clear_title}
\alias{clear_trips}
\title{Clear Animated Arc}
\usage{
clear_animated_arc(map, layer_id = NULL)

clear_line(map, layer_id = NULL)

clear_arc(map, layer_id = NULL)

clear_bitmap(map, layer_id = NULL)

clear_column(map, layer_id = NULL)

clear_geojson(map, layer_id = NULL)

clear_greatcircle(map, layer_id = NULL)

clear_grid(map, layer_id = NULL)

clear_heatmap(map, layer_id = NULL)

clear_hexagon(map, layer_id = NULL)

clear_line(map, layer_id = NULL)

clear_mesh(map, layer_id = NULL)

clear_path(map, layer_id = NULL)

clear_pointcloud(map, layer_id = NULL)

clear_polygon(map, layer_id = NULL)

clear_scatterplot(map, layer_id = NULL)

clear_screengrid(map, layer_id = NULL)

clear_terrain(map, layer_id = NULL)

clear_text(map, layer_id = NULL)

clear_title(map, layer_id = NULL)

clear_trips(map, layer_id = NULL)
}
\arguments{
\item{map}{a mapdeck map object}

\item{layer_id}{the layer_id of the layer you want to clear}
}
\description{
Clears elements from a map

Clears elements from a map
}
