% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{mapdeck}
\alias{mapdeck}
\title{mapdeck}
\usage{
mapdeck(token, data = NULL, width = NULL, height = NULL,
  elementId = NULL, padding = 0,
  style = "mapbox://styles/mapbox/streets-v9", pitch = 0, location = c(0,
  0))
}
\arguments{
\item{token}{Mapbox Acess token}

\item{location}{vector of lon and lat coordinates}
}
\description{
mapdeck
}
