% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{mapdeck}
\alias{mapdeck}
\title{mapdeck}
\usage{
mapdeck(token = get_access_token(api = "mapbox"), data = NULL,
  width = NULL, height = NULL, padding = 0,
  style = "mapbox://styles/mapbox/streets-v9", pitch = 0, zoom = 0,
  location = c(0, 0))
}
\arguments{
\item{token}{Mapbox Acess token. Use \code{set_token()} to set a global token}

\item{data}{data to be used on the map}

\item{width}{the width of the map}

\item{height}{the height of the map}

\item{padding}{the padding of the map}

\item{style}{the style of the map}

\item{pitch}{the pitch angle of the map}

\item{zoom}{zoom level of the map}

\item{location}{vector of lon and lat coordinates (in that order)}
}
\description{
mapdeck
}
