% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_dependencies.R
\name{mapdeck_dependencies}
\alias{mapdeck_dependencies}
\title{Mapdeck Dependencies}
\usage{
mapdeck_dependencies()
}
\description{
Adds the required mapdeck javascript dependencies to a Shiny UI when you want to use
mapdeck layers, but not with a mapdeck map.
}
