% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{mapdeck_legend}
\alias{mapdeck_legend}
\title{Mapdeck Legend}
\usage{
mapdeck_legend(colours, variables, colour_type = c("fill", "stroke"),
  variable_type = c("category", "gradient"), title = "", css = "")
}
\arguments{
\item{colours}{vector of hex colours}

\item{colour_type}{one of "fill" or "stroke"}

\item{variable_type}{one of category (discrete) or gradient (continuous)}

\item{title}{string used as the legend title}

\item{css}{}
}
\description{
Mapdeck Legend
}
