% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{mapdeck_legend}
\alias{mapdeck_legend}
\title{Mapdeck Legend}
\usage{
mapdeck_legend(legend_elements)
}
\arguments{
\item{legend_elements}{vector of legend elements (made from \link{legend_element})}
}
\description{
Constructs legend elements into the correct JSON format for plotting on the map
}
\examples{
l1 <- legend_element(
	variables = c("a","b")
	, colours = c("#00FF00","#FF0000")
	, colour_type = "fill"
	, variable_type = "category"
	, title = "my title"
)

mapdeck_legend(l1)

}
\seealso{
\link{legend_element}
}
