% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{mapdeck_view}
\alias{mapdeck_view}
\title{Mapdeck view}
\usage{
mapdeck_view(map, location, zoom = 6, duration = 0,
  transition = c("linear", "fly"))
}
\arguments{
\item{location}{vector of lon and lat coordinates (in that order)}

\item{zoom}{zoom level of the map}

\item{duration}{time in milliseconds of the transition}

\item{transition}{type of transition}
}
\description{
Changes the view of the of the map
}
