% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapdeck_map.R
\name{mapdeck_view}
\alias{mapdeck_view}
\title{Mapdeck view}
\usage{
mapdeck_view(
  map,
  location = NULL,
  zoom = NULL,
  pitch = NULL,
  bearing = NULL,
  duration = NULL,
  transition = c("linear", "fly")
)
}
\arguments{
\item{map}{a \code{mapdeck} map object}

\item{location}{unnamed vector of lon and lat coordinates (in that order)}

\item{zoom}{zoom level of the map}

\item{pitch}{the pitch angle of the map}

\item{bearing}{bearing of the map between 0 and 360}

\item{duration}{time in milliseconds of the transition}

\item{transition}{type of transition}
}
\description{
Changes the view of the of the map
}
