% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_layer_arc.R
\name{update_arc}
\alias{update_arc}
\title{Update Arc}
\usage{
update_arc(map, data = get_map_data(map), layer_id)
}
\arguments{
\item{layer_id}{the layer_id of the layer to update.}
}
\description{
Update Arc
}
